requirements.txt
streamlit
python-dotenv
pandas
vaderSentiment
scikit-learn
openai
bcrypt
python-multipart
Pillow
